+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 * Prevent the ToO code from selecting the same target for two different positioners.
 * Fix a case when the configuration fails to retrieve the trajectory paths because there are no robots that need updated coordinates.
+* Pin `pymodbus` to 3.7.x.
 
 
 ## 1.8.0 - October 24, 2024